requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
cachetools==3.1.1
cffi==1.13.2
chardet==3.0.4
discord.py==1.2.4
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
multidict==4.5.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyNaCl==1.3.0
rsa==4.0
six==1.13.0
uritemplate==3.0.0
websockets==6.0
yarl==1.3.0
